services/vision/requirements.txt
fastapi==0.109.2
uvicorn[standard]==0.27.1
torch==2.2.0
torchvision==0.17.0
numpy<2
Pillow==10.2.0
httpx==0.26.0
python-multipart==0.0.9
